+++ test/e2e/framework/framework.go
@@ -101,12 +101,6 @@ func (f *Framework) setup() error {
 	}
 	logrus.Info("etcd operator created successfully")
 
-	err = f.SetupEtcdRestoreOperatorService()
-	if err != nil {
-		return err
-	}
-	logrus.Info("etcd restore operator pod and service created successfully")
-
 	logrus.Info("e2e setup successfully")
 	return nil
 }